+++ ReleaseNotes.md
@@ -1,4 +1,14 @@
-# Release Notes
+## 0.8.0-alpha.2
+
+- Update for ubuntu 14.04 and latest docker
+
+- Update to pallet 0.8.0-RC.9
+
+- Update to use docker apt repo for docker 0.7
+  Updates the package source, and fixes the change to :ID from :Id in
+  reported yaml status.
+
+- Update for multiple output format in 0.4.6+
 
 ## 0.8.0-alpha.1